+++ ABPlayer/Sources/Views/Components/VideoProgressView.swift
@@ -4,7 +4,7 @@ import Observation
 // MARK: - Video Progress View
 
 struct VideoProgressView: View {
-  @Environment(AudioPlayerManager.self) private var playerManager
+  @Environment(PlayerManager.self) private var playerManager
 
   @Binding var isSeeking: Bool
   @Binding var seekValue: Double